+++ client/driver/qemu.go
@@ -1,7 +1,9 @@
 package driver
 
 import (
+	"encoding/json"
 	"fmt"
+	"log"
 	"os/exec"
 	"path/filepath"
 	"regexp"
@@ -40,9 +42,11 @@ type QemuDriverConfig struct {
 
 // qemuHandle is returned from Start/Open as a handle to the PID
 type qemuHandle struct {
-	cmd    executor.Executor
-	waitCh chan *cstructs.WaitResult
-	doneCh chan struct{}
+	cmd         executor.Executor
+	killTimeout time.Duration
+	logger      *log.Logger
+	waitCh      chan *cstructs.WaitResult
+	doneCh      chan struct{}
 }
 
 // NewQemuDriver is used to create a new exec driver
@@ -163,7 +167,7 @@ func (d *QemuDriver) Start(ctx *ExecContext, task *structs.Task) (DriverHandle, 
 		if len(forwarding) != 0 {
 			args = append(args,
 				"-netdev",
-				fmt.Sprintf("user,id=user.0%s", strings.Join(forwarding, ",")),
+				fmt.Sprintf("user,id=user.0,%s", strings.Join(forwarding, ",")),
 				"-device", "virtio-net,netdev=user.0",
 			)
 		}
@@ -180,7 +184,8 @@ func (d *QemuDriver) Start(ctx *ExecContext, task *structs.Task) (DriverHandle, 
 	}
 
 	// Setup the command
-	cmd := executor.Command(args[0], args[1:]...)
+	cmd := executor.NewBasicExecutor()
+	executor.SetCommand(cmd, args[0], args[1:])
 	if err := cmd.Limit(task.Resources); err != nil {
 		return nil, fmt.Errorf("failed to constrain resources: %s", err)
 	}
@@ -197,35 +202,58 @@ func (d *QemuDriver) Start(ctx *ExecContext, task *structs.Task) (DriverHandle, 
 
 	// Create and Return Handle
 	h := &qemuHandle{
-		cmd:    cmd,
-		doneCh: make(chan struct{}),
-		waitCh: make(chan *cstructs.WaitResult, 1),
+		cmd:         cmd,
+		killTimeout: d.DriverContext.KillTimeout(task),
+		logger:      d.logger,
+		doneCh:      make(chan struct{}),
+		waitCh:      make(chan *cstructs.WaitResult, 1),
 	}
 
 	go h.run()
 	return h, nil
 }
 
+type qemuId struct {
+	ExecutorId  string
+	KillTimeout time.Duration
+}
+
 func (d *QemuDriver) Open(ctx *ExecContext, handleID string) (DriverHandle, error) {
+	id := &qemuId{}
+	if err := json.Unmarshal([]byte(handleID), id); err != nil {
+		return nil, fmt.Errorf("Failed to parse handle '%s': %v", handleID, err)
+	}
+
 	// Find the process
-	cmd, err := executor.OpenId(handleID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to open ID %v: %v", handleID, err)
+	cmd := executor.NewBasicExecutor()
+	if err := cmd.Open(id.ExecutorId); err != nil {
+		return nil, fmt.Errorf("failed to open ID %v: %v", id.ExecutorId, err)
 	}
 
 	// Return a driver handle
 	h := &execHandle{
-		cmd:    cmd,
-		doneCh: make(chan struct{}),
-		waitCh: make(chan *cstructs.WaitResult, 1),
+		cmd:         cmd,
+		logger:      d.logger,
+		killTimeout: id.KillTimeout,
+		doneCh:      make(chan struct{}),
+		waitCh:      make(chan *cstructs.WaitResult, 1),
 	}
 	go h.run()
 	return h, nil
 }
 
 func (h *qemuHandle) ID() string {
-	id, _ := h.cmd.ID()
-	return id
+	executorId, _ := h.cmd.ID()
+	id := qemuId{
+		ExecutorId:  executorId,
+		KillTimeout: h.killTimeout,
+	}
+
+	data, err := json.Marshal(id)
+	if err != nil {
+		h.logger.Printf("[ERR] driver.qemu: failed to marshal ID to JSON: %s", err)
+	}
+	return string(data)
 }
 
 func (h *qemuHandle) WaitCh() chan *cstructs.WaitResult {
@@ -244,7 +272,7 @@ func (h *qemuHandle) Kill() error {
 	select {
 	case <-h.doneCh:
 		return nil
-	case <-time.After(5 * time.Second):
+	case <-time.After(h.killTimeout):
 		return h.cmd.ForceStop()
 	}
 }